requirements.txt
bert_score==0.3.13
Flask==3.1.3
numpy==2.4.3
protobuf==7.34.0
rank_bm25==0.2.2
rouge_score==0.1.2
scikit_learn==1.8.0
scipy==1.17.1
sentence_transformers==5.1.2
SQLAlchemy==2.0.23
sympy==1.14.0 
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
google-genai